set BASE_CONTAINER as an argument install vim and remove debian stuff
BASE_CONTAINER is set as an argument to docker build similar to the rest
of the jupyter/docker-stacks images. This allows us to do e.g. `make
build/pims-r DARGS="--build-arg BASE_CONTAINER=callysto/scipy-notebook"`
to select a different base image.

The jupyter/base-image sets nano as the editor, but I cry if I can't
find vim, so here it is.

ffmpeg is included in this release of ubuntu so we don't need the
deb-multimedia repository.

Tests pass with minor warnings.

Signed-off-by: Ian Allison <iana@pims.math.ca>